+++ .puppet/modules/grafana/lib/puppet/provider/grafana_dashboard/grafana.rb
@@ -12,9 +12,56 @@ Puppet::Type.type(:grafana_dashboard).provide(:grafana, parent: Puppet::Provider
 
   defaultfor kernel: 'Linux'
 
+  def organization
+    resource[:organization]
+  end
+
+  def grafana_api_path
+    resource[:grafana_api_path]
+  end
+
+  def fetch_organizations
+    response = send_request('GET', format('%s/orgs', resource[:grafana_api_path]))
+    if response.code != '200'
+      raise format('Fail to retrieve organizations (HTTP response: %s/%s)', response.code, response.body)
+    end
+
+    begin
+      fetch_organizations = JSON.parse(response.body)
+
+      fetch_organizations.map { |x| x['id'] }.map do |id|
+        response = send_request 'GET', format('%s/orgs/%s', resource[:grafana_api_path], id)
+        if response.code != '200'
+          raise format('Failed to retrieve organization %d (HTTP response: %s/%s)', id, response.code, response.body)
+        end
+
+        fetch_organization = JSON.parse(response.body)
+
+        {
+          id: fetch_organization['id'],
+          name: fetch_organization['name']
+        }
+      end
+    rescue JSON::ParserError
+      raise format('Failed to parse response: %s', response.body)
+    end
+  end
+
+  def fetch_organization
+    unless @fetch_organization
+      @fetch_organization =
+        if resource[:organization].is_a?(Numeric) || resource[:organization].match(%r{^[0-9]*$})
+          fetch_organizations.find { |x| x[:id] == resource[:organization] }
+        else
+          fetch_organizations.find { |x| x[:name] == resource[:organization] }
+        end
+    end
+    @fetch_organization
+  end
+
   # Return the list of dashboards
   def dashboards
-    response = send_request('GET', '/api/search', nil, q: '', starred: false)
+    response = send_request('GET', format('%s/search', resource[:grafana_api_path]), nil, q: '', starred: false)
     if response.code != '200'
       raise format('Fail to retrieve the dashboards (HTTP response: %s/%s)', response.code, response.body)
     end
@@ -30,7 +77,7 @@ Puppet::Type.type(:grafana_dashboard).provide(:grafana, parent: Puppet::Provider
   def find_dashboard
     return unless dashboards.find { |x| x['title'] == resource[:title] }
 
-    response = send_request('GET', format('/api/dashboards/db/%s', slug))
+    response = send_request('GET', format('%s/dashboards/db/%s', resource[:grafana_api_path], slug))
     if response.code != '200'
       raise format('Fail to retrieve dashboard %s (HTTP response: %s/%s)', resource[:title], response.code, response.body)
     end
@@ -44,6 +91,12 @@ Puppet::Type.type(:grafana_dashboard).provide(:grafana, parent: Puppet::Provider
   end
 
   def save_dashboard(dashboard)
+    # change organizations
+    response = send_request 'POST', format('%s/user/using/%s', resource[:grafana_api_path], fetch_organization[:id])
+    unless response.code == '200'
+      raise format('Failed to switch to org %s (HTTP response: %s/%s)', fetch_organization[:id], response.code, response.body)
+    end
+
     data = {
       dashboard: dashboard.merge('title' => resource[:title],
                                  'id' => @dashboard ? @dashboard['id'] : nil,
@@ -51,8 +104,8 @@ Puppet::Type.type(:grafana_dashboard).provide(:grafana, parent: Puppet::Provider
       overwrite: !@dashboard.nil?
     }
 
-    response = send_request('POST', '/api/dashboards/db', data)
-    return unless response.code != '200'
+    response = send_request('POST', format('%s/dashboards/db', resource[:grafana_api_path]), data)
+    return unless (response.code != '200') && (response.code != '412')
     raise format('Fail to save dashboard %s (HTTP response: %s/%s', resource[:name], response.code, response.body)
   end
 
@@ -73,7 +126,7 @@ Puppet::Type.type(:grafana_dashboard).provide(:grafana, parent: Puppet::Provider
   end
 
   def destroy
-    response = send_request('DELETE', format('/api/dashboards/db/%s', slug))
+    response = send_request('DELETE', format('%s/dashboards/db/%s', resource[:grafana_api_path], slug))
 
     return unless response.code != '200'
     raise Puppet::Error, format('Failed to delete dashboard %s (HTTP response: %s/%s', resource[:title], response.code, response.body)