+++ .puppet/modules/grafana/lib/puppet/type/grafana_plugin.rb
@@ -5,6 +5,12 @@ manages grafana plugins
 @example Install a grafana plugin
  grafana_plugin { 'grafana-simple-json-datasource': }
 
+@example Install a grafana plugin from different repo
+ grafana_plugin { 'grafana-simple-json-datasource':
+   ensure => present,
+   repo   => 'https://nexus.company.com/grafana/plugins',
+ }
+
 @example Uninstall a grafana plugin
  grafana_plugin { 'grafana-simple-json-datasource':
    ensure => absent,
@@ -27,5 +33,15 @@ DESC
   newparam(:name, namevar: true) do
     desc 'The name of the plugin to enable'
     newvalues(%r{^\S+$})
+  end
+
+  newparam(:repo) do
+    desc 'The URL of an internal plugin server'
+
+    validate do |value|
+      unless value =~ %r{^https?://}
+        raise ArgumentError, format('%s is not a valid URL', value)
+      end
+    end
   end
 end