+++ .puppet/modules/grafana/spec/spec_helper_acceptance.rb
@@ -1,4 +1,5 @@
 require 'beaker-rspec'
+require 'beaker-puppet'
 require 'beaker/puppet_install_helper'
 require 'beaker/module_install_helper'
 
@@ -7,7 +8,7 @@ install_module
 install_module_dependencies
 
 # Install additional modules for soft deps
-install_module_from_forge('puppetlabs-apt', '>= 4.1.0 < 5.0.0')
+install_module_from_forge('puppetlabs-apt', '>= 4.1.0 < 7.0.0')
 
 RSpec.configure do |c|
   # Readable test descriptions